Added more functions for analyzing spatial data
