+++ cell2cell/spatial/__init__.py
@@ -1,3 +1,3 @@
 from cell2cell.spatial.distances import (celltype_pair_distance, pairwise_celltype_distances)
-from cell2cell.spatial.neighborhoods import (create_spatial_grid, dist_filter_liana, dist_filter_tensor)
-
+from cell2cell.spatial.filtering import (dist_filter_liana, dist_filter_tensor)
+from cell2cell.spatial.neighborhoods import (create_spatial_grid, create_moving_windows, calculate_window_size, add_moving_window_info_to_adata)
