+++ release/0.7.0-notes.md
@@ -2,7 +2,11 @@
 
 ## New features
 - Added `cell2cell.spatial` functions for enabling analyses considering spatial organization in spatial data.
- Basic analyses are implemented, more complex methods will be implemented in the future.
+These functions include:
+  - Filtering by intercellular distances by thresholding values (`cell2cell.spatial.distances` and `cell2cell.spatial.filtering`).
+  - Dividing the tissue in square grids (`cell2cell.spatial.neighborhoods.create_spatial_grid()`)
+  - Dividing the tissue in moving windows (`cell2cell.spatial.neighborhoods.create_moving_windows()`, `cell2cell.spatial.neighborhoods.calculate_window_size()`,
+    and `cell2cell.spatial.neighborhoods.add_moving_window_info_to_adatae()`)
 
 ## Feature updates
